pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling", "uv-dynamic-versioning" ]

[project]
name = "spatial-collection"
description = "spatialcollection"
readme = "README.md"
license = { file = "LICENSE" }
maintainers = [
  { name = "Sricharan Reddy Varra", email = "sricharanvarra@gmail.com" },
]
authors = [ { name = "Sricharan Reddy Varra" } ]
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dynamic = [ "version" ]
dependencies = [
  # for debug logging (referenced from the issue template)
  "boltons>=25",
  "emoji>=2.14.1",
  "loguru>=0.7.3",
  "more-itertools>=10.6",
  "narwhals>=1.35",
  "numcodecs<0.16",
  "patito>=0.8.3",
  "polars>=1.27.1",
  "pydantic>=2.11.3",
  "pydantic-zarr>=0.7",
  "rich>=14",
  "richuru>=0.1.1",
  "session-info2",
  "spatialdata>=0.3.1.dev31",
  "universal-pathlib>=0.2.6",
]
optional-dependencies.dev = [ "esbonio>=0.16.5", "pre-commit" ]
optional-dependencies.doc = [
  "docutils>=0.8,!=0.18.*,!=0.19.*",
  "ipykernel",
  "ipython",
  "myst-nb>=1.1",
  "pandas",
  # Until pybtex >0.24.0 releases: https://bitbucket.org/pybtex-devs/pybtex/issues/169/
  "setuptools",
  "sphinx>=4",
  "sphinx-autodoc-typehints",
  "sphinx-book-theme>=1",
  "sphinx-copybutton",
  "sphinx-tabs",
  "sphinxcontrib-bibtex>=1",
  "sphinxext-opengraph",
]
optional-dependencies.nb = [ "jupyterlab>=4.4" ]
optional-dependencies.test = [ "coverage", "pytest" ]
# https://docs.pypi.org/project_metadata/#project-urls
urls.Documentation = "https://spatial-collection.readthedocs.io/"
urls.Homepage = "https://github.com/srivarra/spatial-collection"
urls.Source = "https://github.com/srivarra/spatial-collection"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.pytest.ini_options]
testpaths = [ "tests" ]
xfail_strict = true
addopts = [
  "--import-mode=importlib", # allow using test files with same name
]

[tool.coverage.run]
source = [ "spatialcollection" ]
omit = [ "**/test_*.py" ]

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = true

[tool.cruft]
skip = [
  "tests",
  "src/**/__init__.py",
  "src/**/basic.py",
  "docs/api.md",
  "docs/changelog.md",
  "docs/references.bib",
  "docs/references.md",
  "docs/notebooks/example.ipynb",
]

[tool.uv.sources]
spatialdata = { git = "https://github.com/scverse/spatialdata", branch = "main" }
